+++ build.sh
@@ -1,5 +1,3 @@
-#!/bin/bash
-
 # Update version path
 VERSION_FILE="VERSION"
 current_version=$(grep -E "^VERSION=" "$VERSION_FILE" | cut -d'=' -f2)
@@ -9,15 +7,18 @@ new_version="$major.$minor.$patch"
 sed -i "s/^VERSION=.*/VERSION=$new_version/" "$VERSION_FILE"
 echo "Updated version to $new_version"
 
-# Set package version
 source VERSION
-sed -e 's#{VERSION}#'"${VERSION}"'#g' setup_template.py > setup.py
+# Detecta branch atual e adiciona -beta se não for main/master
+CURRENT_BRANCH=$(git rev-parse --abbrev-ref HEAD)
+if [[ "$CURRENT_BRANCH" != "main" && "$CURRENT_BRANCH" != "master" ]]; then
+  VERSION="${VERSION}-b.0"
+fi
 
-# Build package
+sed -e 's#{VERSION}#'"${VERSION}"'#g' pyproject_template.toml > pyproject.toml
+
 rm -R build/
-python3 setup.py build sdist bdist_wheel
+poetry build
 
-# Commit as a tag on repository
 git add --all
 git commit -m "Building a new version ${VERSION}"
 git tag -a ${VERSION} -m "Building a new version ${VERSION}"